Add verbose logging and timestamp to common logger
