+++ Samples/EdgeModules/SqueezeNetObjectDetection/cs/SqueezeNetObjectDetection.csproj
@@ -17,6 +17,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\Common\CS\Common\Common.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
     <PackageReference Include="Mono.Options.Core" Version="1.0.0" />
     <PackageReference Include="Microsoft.Azure.Devices.Client" Version="1.19.0" />
